backend/requirements.txt
flask
openai
python-dotenv
flask-cors
gunicorn
